+++ src/main/resources/schema.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS `performance`
 (
     `id`         BINARY(16) default (uuid())              NOT NULL COMMENT '공연/전시 ID',
     `name`       varchar(255)                             NOT NULL COMMENT '공연/전시 이름',
-    `price`      INT                                      NOT NULL COMMENT '가격',
+    `price`      BIGINT                                   NOT NULL COMMENT '가격',
     `round`      INT                                      NOT NULL COMMENT '회차',
     `type`       INT                                      NOT NULL COMMENT 'NONE, CONCERT, EXHIBITION',
     `start_date` datetime                                 NOT NULL COMMENT '공연 일시',
